Atualizações novas

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,12 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
-using ProEventos.Persistence.Contexto;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -25,7 +22,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventosService.GetAllEventosASync(true);
-                if(eventos == null) return NotFound("Nenhum Evento encontrado.");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -43,7 +40,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventosService.GetEventoByIdASync(id, true);
-                if(evento == null) return NotFound("Evento por Id não encontrado.");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -61,7 +58,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventosService.GetAllEventosByTemaASync(tema, true);
-                if(evento == null) return NotFound("Eventos por tema não encontrados.");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -74,12 +71,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventosService.AddEventos(model);
-                if(evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -92,12 +89,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await _eventosService.UpdateEventos(id, model);
-                if(evento == null) return BadRequest("Erro ao tentar atualizar evento.");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -114,9 +111,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventosService.GetEventoByIdASync(id, true);
+                if(evento == null) return NoContent();
+
                 return await _eventosService.DeleteEvento(id) ?
                     Ok("Deletado") :
-                    BadRequest("Evento não deletado");
+                    throw new Exception("Ocorreu um problema ao deletar o evento.");
             }
             catch (Exception ex)
             {
